requirements.txt
sentence-transformers==2.2.2
psycopg2-binary==2.9.7
torch>=1.9.0
transformers>=4.21.0
huggingface-hub==0.16.4
